refactor: move task details files to a flat structure
