+++ tasklist/client/src/v1/TaskDetails/Variables/JSONEditorModal/index.tsx
