+++ tasklist/client/src/v1/TaskDetails/Variables/JSONEditorModal/styles.module.scss
