+++ tasklist/client/src/v1/TaskDetails/Variables/ResetForm.tsx
