+++ tasklist/client/src/v1/TaskDetails/Variables/VariableEditor/LoadingTextarea/styles.module.scss
