+++ tasklist/client/src/v1/TaskDetails/Variables/VariableEditor/OnNewVariableAdded.tsx
