+++ tasklist/client/src/v1/TaskDetails/Variables/VariableEditor/styles.module.scss
